requirements.txt
flask
requests
python-chess
cairosvg
